changed class PointList definition
